Remove files of packages not present in repodata
This requires that all relevant repodata for an architecture be passed
on the command line at startup, and that the cache be localized to just
that architecture, right now. This isn't too big of a deal, but is
mildly annoying. However, without the ability to preserve a more
accurate DB of things seen and access that from multiple processes,
it's risky to do much else there.

This differs from sample .vmandump files in that the structure is
slightly modified to not be compatible with existing ones because the
current ones all have absolute paths. xmandump expects all caches to
have relative paths. basically, a cache file controls its directory's
copy of the files.

In adddition, there are a few other changes:

- Disable log sampling and ergo rate limiting. This makes it hard to
  debug and doesn't benefit anyone when it's not a daemon.

- Record packages without manpages as empty arrays. This was causing
  some annoying reprocessing of package files because I neglected to
  handle the case of packages that were already scanned but just not in
  the cache.