Scan packages' files.plist before scanning tar files
This speeds up package scanning enough that where it used to take 13
minutes to scan my build server, it only takes 40-50 seconds now.

It works roughly by taking the files.plist, scanning it for manpage
dirs, and if manpage dirs are present, then scanning its files list for
all files under the manpage dirs. It then walks to rest of the tar file
looking for those manpages. Once all of them have been found and
extracted, it's done with the tar file and closes it.

So, where it takes about four minutes to scan e.g., ghc or ghc-bin,
with this naive improvement, it takes about 11s.

There are some other improvements that can likely be made, but I doubt
anything will be as significant as this.

In addition, after each package scan, the code now forces a GC. This is
because, without this, the can and will use a gigabyte or more of
memory, and the GC won't really do much about it until it grows fairly
large.